+++ src/declarative/components/pwm/VibrationMotorModule.ts
