+++ src/index.ts
@@ -15,8 +15,8 @@ export { Servo } from './declarative/components/servo/Servo'
 export { DigitalTileSensor } from './declarative/components/input/DigitalTileSensor'
 export { InfraredObstacleAvoidance } from './declarative/components/input/InfraredObstacleAvoidance'
 export { PhotoInterrupter } from './declarative/components/input/PhotoInterrupter'
-export { VibrationMotorModule } from './declarative/components/PWM/VibrationMotorModule'
+export { VibrationMotorModule } from './declarative/components/pwm/VibrationMotorModule'
 export { Output } from './declarative/components/output/Output'
-export { RGBLed } from './declarative/components/PWM/RGBLed'
+export { RGBLed } from './declarative/components/pwm/RGBLed'
 
 export type { SerialPort } from 'serialport'
